tetragon: Add traceBenchCustom bench
Allows to run custom benchmark with specific crd file, like:

  # ./tetragon-bench -crd crds/examples/names.yaml  -- perf bench sched pipe

Or with perf sampling:

  # perf record -g -e cycles:k ./tetragon-bench -crd ./read_write.yaml -json-encode -- perf bench sched pipe

Arguments after '--' will be executed as workload. When workload
is finished the bench will exit.

+++ pkg/bench/trace.go
@@ -17,7 +17,7 @@ type TraceBench interface {
 }
 
 var (
-	traceBenches = []string{"rw", "open"}
+	traceBenches = []string{"rw", "open", "custom"}
 )
 
 func TraceBenchSupported() []string {
@@ -55,12 +55,17 @@ func RunTraceBench(args *Arguments) (summary *Summary) {
 		bench = newTraceBenchRw()
 	case "open":
 		bench = newTraceBenchOpen()
+	case "custom":
+		bench = newTraceBenchCustom()
 	default:
 		panic("unknown benchmark")
 	}
 
 	configFile := bench.ConfigFilename(ctx, args)
-	defer os.Remove(configFile)
+
+	if args.Trace != "custom" {
+		defer os.Remove(configFile)
+	}
 
 	// Start tetragon if requested.
 	tetragonFinished := make(chan bool, 1)
